Bug #345025 phoneserver: fix 64-bit runtime error
[bug number  ] 345025
[root cause  ]
[changes     ]
[side effects]
[self test   ] 7731: enter android
[whether AOB ]
[reviewers   ] brady

+++ phoneserver/cmux.h
@@ -47,7 +47,7 @@ struct cmux_t {
 	int cmd_type;		/*## cmd  type */
 	int wait_resp;		/*## flag for wait_resp */
 	int in_use;
-	int (*callback) (AT_CMD_RSP_T * resp_req, int usdata);
+	int (*callback) (AT_CMD_RSP_T * resp_req, unsigned long usdata);
 	char name[30];
 	struct cmux_ops *ops;
 	sem cmux_lock;
